Undo plugins merge
